test/autotest.sh: properly name test case/suite
There hasn't been much of a semantic distinction between case and suite
up to this point. A suite should contain multiple tests. A suite can
contain multiple commands or multiple steps, but it should not contain
any nested calls to run_test.

This is the first of several changes that make that clarification across
the code base.

Change-Id: Ie3c428932a41f172a6f114c69ee439db377f9853
Signed-off-by: Seth Howell <seth.howell@intel.com>