bdev/ocssd: wrap IO part of bdev_ocssd_io
Wrap the part of the bdev_ocssd_io responsible for describing IO
operations in a separate structure.  Currently all fields are wrapped
this way, but it'll allow the bdev_ocssd_io to become a union of several
different types of structures (one for the IO operations, one for
getting zone information, etc.).

Change-Id: I72276ce786d3c9ef0d1762c5ac20071b7a163ef5
Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>