Update patch set 1
Patch Set 1:

For async connection construction, we need to decouple the connection construction and call the connection completion separately, and spdk nvme-oF initiator can be used in different applications, so we cannot use poller. 

And this patch can be a temporal solution. Let's have more discussion later to make how to change small code to support the async connection from the remote NVMe-oF target if used in spdk bdev layer based on SPDK framework.

Patch-set: 1