vhost: always set VHOST_USER_PROTOCOL_F_CONFIG protocol feature bit
VHOST_USER_PROTOCOL_F_CONFIG is defined for all vhost user devices not
only for vhost_blk, DPDK doesn't want to maintain it, so we implemented
it here.  The macro SPDK_CONFIG_VHOST_INTERNAL_LIB will be only for
vhost_nvme, so we can always enable this protocol feature bit.

Change-Id: I9f8d8bcf3652e6c3a71a7ef04b9af66082857fd1
Signed-off-by: Changpeng Liu <changpeng.liu@intel.com>