lib/bdev: Remove duplicate call to bdev_io_put_buf()
When we unset a bounce buffer we call bdev_io_put_buf() to put it
back to the right pool however we also do this in spdk_bdev_free_io()
as the same pointer is used for a bounce allocation or if the caller
didn't provide one.

Does not hurt anything, just working in this area on something else
and felt like it was worth minor cleanup.  Also fixed a comment typo
in the same function.

Signed-off-by: paul luse <paul.e.luse@intel.com>
Change-Id: Id14b49e8b7bc2372ecb8dd7579e6797a395c2cf2