Update patch set 8
Patch Set 8: Code-Review-1

This is going to significantly reduce the performance of blobfs. The purpose of blobfs is to be a fast, lightweight alternative to a real filesystem. It is not intended to be a full filesystem replacement. At a minimum, capturing time stamps needs to be optional, but I question whether doing this makes sense at all.

Patch-set: 8
Reviewer: Gerrit User 1010531 <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1