Update patch set 6
Patch Set 6:

> Patch Set 6: Code-Review-1
> 
> I've ran nightly on patch next in series to verify previous ones (otherwise I'd get errors on hello_bdev).
> https://dqtibwqq6s6ux.cloudfront.net/public_build/autotest-nightly_1155.html
> It seems like sourcing the scripts instead of executing them changes the behavior of 'shell_restore_x' in vm_setup().

Thanks so much for catching this! Actually I am still sourcing the files, so that wasn't the problem turns out that I actually ended up calling xtrace_disable twice in a row in run_test which left x unset when assigning shell_restore_x. if x is unset when shell_restore_x is assigned, it returns an error code and the script fails. So I added a couple of changes. One to make xtrace_disable capable of handling multiple calls without losing the original behavior and a second to completely remove shell_restore_x which was just trying to do the same thing as xtrace_disable.

Patch-set: 6