Update patch set 1
Patch Set 1:

I'm 99.9% percent sure that we don't want to respect this SC.
The reason is that our scripts depend on each other via sourcing.

A variable may exist in one script and not be used in it, but will be used as a "global" variable in other script. Typical example is the use of common.sh files for tests.

Patch-set: 1
Reviewer: Gerrit User 1011256 <1011256@d5d70762-12d0-45a1-890d-524b12d3f735>