Update patch set 7
Patch Set 7:

> Patch Set 7: Code-Review-1
> 
> Vhost migration seem to fail on this patch.
> Manual nightly test run: https://10.102.17.104:8080/job/Vhost_tests/job/vhost_autotest_nightly/4505/consoleFull
> 
> Will try to dig deep into that tomorrow.

Hi Karol, thanks for the heads up on that. It looks like it was related to a syntax problem in vm_setup that uncovered this issue when I removed the --test-cases option from the tests. It's unrelated to this patch, so I put it in its own patch off master.

https://review.gerrithub.io/c/spdk/spdk/+/478079

Here's the proof it's passing on this patch after the fix. https://10.102.17.104:8080/view/all/job/Vhost_tests/job/vhost_autotest_nightly/4516/

Patch-set: 7