lib/reduce: fix zero_buff alloc/dealloc
The g_vol_count value was not being incremented during volume load in
the destroy path resulting in the volume count falling to -1.
When we call _load_read_super_and_path_cpl, we do an explicit check to
see if we are in the destroy path, and if we are, we short circuit to
destroying the reduce volume before we  ever get a chance to call
_alloc_zero_buff.
So when we have a single volume and we call spdk_reduce_vol_unload and
then spdk_reduce_vol_destroy, we will end up decrementing the volume
count twice for a volume without doing an increment in between.
Make sure we always increment g_vol_count on loads, not just when
we get to the point where we allocate the zero buffer.

Change-Id: Iaaf2df0659d345ea8fd45be7e08de749bb5468f7
Signed-off-by: Seth Howell <seth.howell@intel.com>