lib/reduce: fix zero_buff alloc/dealloc
The g_vol_count value was not being incremented during volume load in
the destroy path resulting in the volume count falling to -1.
So add an extra NULL check into the path where we free the buffer so
that we don't decrement past zero.
Note, this also protects against the case where an app is misusing the
library and attempts multiple frees on the zero buffer using public api
calls to spdk_reduce_vol_unload.

Change-Id: Iaaf2df0659d345ea8fd45be7e08de749bb5468f7
Signed-off-by: Seth Howell <seth.howell@intel.com>