test/bdev: remove the gpt conf file
This file wasn't actually doing what we thought. If you look at the code
in part_dev_by_gpt, we were setting Disable Yes in the configuration
file. If you look at the code in vbdev_gpt.c, when Disable Yes is set,
that actually sets the global flag g_gpt_disabled to false which means
that the gpt configuration will get done in that case.

So long story short this configuration file isn't needed and we can just
get rid of it.

Change-Id: Ia6ded22ce16fc1f76b7d99643b9d37e3ecbd1c60
Signed-off-by: Seth Howell <seth.howell@intel.com>