Update patch set 3
Patch Set 3: Code-Review-1

(1 comment)

So the way this works is that a heap contains allocations for memzones and raw allocations (i.e. spdk_malloc and variants).  The mempool contests of three memzones.

When we show the total memory size, it's for the heaps.  The mempools and memzones are already part of that.

I think this is important, because we don't want to make it seem like there's 500MB allocated in the heap, and then another 450MB in mempools.  It's just the 500MB.

You could be *really* slick, and parse the addresses of the memzones, and match those with each of the malloc elements.  Then whichever malloc elements are remaining, you know are from spdk_malloc.  Such as a malloc bdev.

But for now, I'd maybe word it something like this:

DPDK memory size 672.000000 MiB in 1 heap(s)

heaps totaling size 672.000000 MiB
  size: 672.000000 MiB heap id: 0

mempools totaling size 624.674500 MiB
  size: 200.758362 MiB name: PDU_immediate_data_Pool heap:0

memzones totaling size xxxx MiB
  size: 1.000366 MiB name: RG_ring_0_12650 heap:0

Patch-set: 3
Reviewer: Gerrit User 1011222 <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1