lib/bdev: Add support for new per bdev_io auxiliary buffer.
New elements were added to the `bdev_io` structure allowing the caller to request
an auxiliary buffer for its own private use.  The caller requests the buffer by first
setting the `need_aux_buf` flag to `true` then calling `spdk_bdev_io_get_buf(). The
buffer will be described by `aux_buf` and `aux_buf_len`.

The initial use case is crypto, in the next patch in series it is used. No UT were
added as the logic isn't that complicated and it is fully tested with each run
of crypto.

Signed-off-by: paul luse <paul.e.luse@intel.com>
Change-Id: Ib1939fcbc8e5db36fd909ef26771a725a551e8e6