Update patch set 8
Patch Set 8:

Thanks for your review, Paul.

> Is there any other benefit from this other than saving a potential malloc failure? Guess I'm not sure why we need to do this here.

Definition of struct spdk_cpuset was concealed before, so spdk_cpuset_alloc() was the only way to use cpumask.

But generally struct cpuset is exposed publicly and so I added a change to move the definition to the header file a while ago.

So this change is possible now.

This change not only avoids malloc failure but also simplifies the code.

spdk_cpuset_alloc() may be deprecated but that is another topic.

Patch-set: 8