[RFC] nvmf: add new muser transport
This is an aggregated diff of the muser changes in the
Nutanix muser repository found at:

https://github.com/tmakatos/spdk.git

This patch is specifically for collecting initial feedback.
Nutanix will eventually push this functionality with a
different patch once ready.

This specific patch is a diff of commits:

92df9955 from SPDK master branch:
(json: increase the json rpc client value limit)

635ef943 from tmakatos master branch:
(remove obsolete check regarding lazy map of admin queues)

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: Iaf51d71f0f734cd2bd4e779c673a6ad24b756649