Update patch set 18
Patch Set 18:

Seth and I changed CI to allow some of the 1G VMs to be used for general purpose tests (not just ghost-initiator).  But libfuse wasn't installed on those systems which caused this test failure.

Seth is getting that fixed up.  We're not going to file a GitHub issue for this.

Patch-set: 18
Reviewer: Gerrit User 1011222 <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>