bdevperf: Call poller_unregister() on the correct thread for LUN hotplug
We did not check if spdk_poller_unregister is called() on the thread
which called spdk_poller_register().  spdk_poller_unregister() had
been called on the master thread but it had to be called on the
thread on which target is scheduled.

Move two poller_unregister() calls from bdevperf_target_gone() to
bdevperf_complete() to call poller_unregister() on the correct thread.

Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>
Change-Id: I7db266a74679008f604f95fcfeb92655944b25a8