bdevperf: Factor out unregistering poller and enabling is_draining into a function
bdevperf_target_gone() and end_target() have the same operations,
and so factor out contents of them into a helper function _end_target().
_end_target() doesn't need to have two parameters but this is needed to
use event call in the next patch.

bdevperf_target_gone() has a bug now and it is that poller_unregister()
has to be called on the thread that runs the coresponding target.

This factorization makes us easier to use event call to fix the bug and
we will do it in the next patch.

Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>
Change-Id: Ie1f7120ea3779580626ebdbdf1d9c17d99a4f42a