nvme: Don't use stack variable to track request completion
A pointer to a stack variable is passed as an argument to
nvme_completion_poll_cb function, later this variable is used
to track completion in the spdk_nvme_wait_for_completion() function.
If normal scenario a request submitted to the admin queue will be completed
within the function which submitted the request.
spdk_nvme_wait_for_completion() calls nvme_transport_qpair_process_completions
which may return an error to the caller, the caller may exit from the
function which submitted the request and the pointer to the stack variable
will no longer be valid. Thereby the request may not be completed at that time
and completed later (e.g. when the controller/qpair are destroyed)
and that will lead to call to nvme_completion_poll_cb with the pointer
to invalid stack variable.
Fix - add a new field of type nvme_completion_poll_status to spdk_nvme_ctrlr
which is used to track completion of a request.

Fixes #1125

Change-Id: I8f04e79e9f767f9f938c68536eb5f1d6974f4b63
Signed-off-by: Alexey Marchuk <alexeymar@mellanox.com>