Update patch set 8
Patch Set 8:

Jim suggests that we create two new API names for this. With this patch, we can compile several sock implementations together. Then the applications can use different sock implementations in one application together. Definitely, when we use RPC or configuration file, we need add a hint in the ip address. For example, Listen TCP 10.67.110.181:4420 -> Listen TCP 10.67.110.181:4420 posix. 

The reason that we need this patch, for the same 10.67.110.181:4420 address (for example), several implementations can both open/connect on the address. Then we will pick up the first implementation which can listen/connect on the address, which is not flexible.

Patch-set: 8