lib/blob: pass Extent Page offset on cluster allocation
Extent Pages claim and insertion can be asynchronous
in case of cluster allocation by writing to new cluster.

In that case a cluster and extent page is first claimed,
then message is passed to md_thread to actually insert
them into the md and sync.

This patch adds parameters to pass the Extent Page offset
in such case.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Change-Id: I46d8ace9cd5abc0bfe48174c2f2ec218145b9c75