Update patch set 4
Patch Set 4:

> Patch Set 3: Code-Review-1
> 
> Should we translate the sc into an errno?  For example, I think in most of these cases, if the cpl indicates an error, we'll want to return EIO (which is 5).
> 
> Things like SC_INVALID_FIELD (0x2) would get translated into ENOENT (No File or Directory) which would be really confusing.
> 
> In fact, for now, I'd say just always return EIO if the cpl indicates an error.  You can use:
> 
> spdk_nvme_cpl_is_error(cpl) ? 0 : EIO

Linux kernel driver returns sc here or negative number for linux errors. nvme-cli translates it to the errno itself and provides descriptive status information in command output, so changing it will have impact on the behavior of commands like nvme-cli which expects it here (e.g. returned error codes, status information and so on).

I think that it's better to not translate it.

Patch-set: 4