Update patch set 10
Patch Set 10: Code-Review-1

(1 comment)

I think we need to go carefully through the requirements here. I'm not yet convinced that applications need to be able to use two different implementations simultaneously.

Certainly, the libraries need to be able to have multiple implementations available to them when they're compiled. The selection of implementation needs to be a run-time decision, instead of compile-time like it is today. But that doesn't mean that they need to select the implementation on a per-socket basis.

Patch-set: 10
Reviewer: Gerrit User 1010531 <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1