[RFC] nvmf: add new muser transport
This is an aggregated diff of the muser changes in the
Nutanix muser repository found at:

https://github.com/tmakatos/spdk.git

This patch is specifically for collecting initial feedback.
Nutanix will eventually push this functionality with a
different patch once ready.

This specific patch is a diff of commits:

801e76d8b from SPDK master branch:
(bdevperf: Change g_run_failed to g_run_rc and use it correctly)

635ef943 from tmakatos master branch:
(remove obsolete check regarding lazy map of admin queues)

Comments addressed since patchset 2:
  * remove obsolete SQ0TBDL, CQ0HDBL, and spdk_nvme_dptr
  * remove subnqn_t and use SPDK_NVME_NQN_FIELD_SIZE instead
  * remove unnecessary zeroing of memory
  * used cdw10/11 for creating/deleting I/O queue from nvme_spec.h
  * fixed size of BA in MLBAR

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Signed-off-by: Thanos Makatos <thanos.makatos@nutanix.com>
Change-Id: Iaf51d71f0f734cd2bd4e779c673a6ad24b756649