Update patch set 10
Patch Set 10:

(1 comment)

I think providing a way to specify the impl would be helpful for testing purposes.  I do not think that a service provider would purposely select posix for example if io_uring was available though.  I think they would just do throttling instead.  Otherwise the service provider ends up paying for a less efficient interface.

There are a few alternatives here:

1) Break the sock API and add a const char * parameter to name the impl to the existing listen/connect calls.  If user passes NULL, then it just uses the default behavior (tries each impl in order).
2) Add a brand new API here like Ziye does with this patch.
3) Similar to #1, but add the const char * parameter as a vararg, so that the existing API is not broken.
4) Provide a separate API that specifies the priority order of the impls.  Users would need to set the priority before creating any kind of listen socket or a connection.

I'm leaning towards #1.  I think users of this API are very minimal at the moment, and changing the API here is easily workaround around (just add NULL as a parameter).

I think #3 is overkill.  #2 would be fine, but if we're OK with breaking the API then I think #1 is clearly better.

I think we still may need some aspects of #4 to specify the default order.  Or we need to add parameters to SPDK_NET_IMPL_REGISTER() to specify relative priorities - for example, make io_uring the highest priority if it is available.  This wasn't an issue with vpp (or Seastar) since in those cases, a given IP address would only be valid for one specific impl.  But now with multiple impls for the kernel TCP stack, we will want to specify an explicit priority order and not rely on the order in which they are registered.

Patch-set: 10
Reviewer: Gerrit User 1011222 <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>