Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hi Gang,
> 
> Reactor knows what core each thread runs on in the context of thread, lw_thread.
> This looks related with what I discussed about this in the last community meeting.
> https://trello.com/c/ZnO7OGis/265-how-spdk-thread-calls-functions-specific-to-reactor
> 
> Adding socket ID to output of framework_get_reactors will satisfy your need?
> Otherwise reactor have to register its callback to thread library to notify this information.
> 
> -Shuhei

Looks good to me and let me add the socket information there.

The major purpose is that for the reactor and its lw_thread, there is socket information and especially for the NVMe Bdev, there is also socket information. Trying to have some message on whether both of them are from the same socket or not.

Patch-set: 1