Update patch set 9
Patch Set 9:

> Patch Set 9: Code-Review-1
> 
> Would you be OK if we only supported RPC for zone_block?
> 
> We want to eventually deprecate the old config file format, and only use JSON-RPC format.  If you think this is required, I could be convinced to add it, but my preference would be to leave it out if it's not really required.  I think we've actually removed most (if not all) of the text config file usage from all of our automated tests.

This patch could be definitely removed. We only added config file support to make FTL and this bdev testable with fio plugin, which didn't support json. If the support was added already (last time we checked it was under review I believe), then this change can certainly be dropped.

Patch-set: 9